% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebsmoothr-package.R
\docType{package}
\name{ebsmoothr-package}
\alias{ebsmoothr}
\alias{ebsmoothr-package}
\title{ebsmoothr: Empirical Bayes Smoothing}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
