% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebsmoothr-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{nbinom2}
\alias{betabinomial}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{glmmTMB}{\code{\link[glmmTMB:nbinom2]{betabinomial}}, \code{\link[glmmTMB]{nbinom2}}}
}}

