% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb_smoother.R
\name{eb_smoother}
\alias{eb_smoother}
\title{Empirical Bayes Smoother}
\usage{
eb_smoother(data, observed, population, family = nbinom2(), formula = ~1, ...)
}
\arguments{
\item{data}{A data frame.}

\item{observed}{A variable representing the observed counts specified in the
same way as \code{\link[dplyr:pull]{dplyr::pull()}}.}

\item{population}{A variable representing the population counts specified in
the same way as \code{\link[dplyr:pull]{dplyr::pull()}}.}

\item{family}{A family object created by \code{\link[=nbinom2]{nbinom2()}} or \code{\link[=betabinomial]{betabinomial()}}.}

\item{formula}{A one-sided formula specifying the model.}

\item{...}{Additional arguments passed to \code{\link[glmmTMB:glmmTMB]{glmmTMB::glmmTMB()}}.}
}
\value{
An object of class \code{eb_smoother}.
}
\description{
Empirical Bayes Smoother
}
